lazy route完全に理解した
